Simplify `react-web-components`
