--- examples/react-web-components/.eslintignore
@@ -1,1 +1,0 @@
-dist
