--- examples/react-web-components/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-.cache
-public
