+++ examples/react-web-components/package.json
@@ -3,15 +3,15 @@
   "version": "2.0.0-next.9",
   "private": true,
   "scripts": {
-    "build": "gatsby build",
-    "start": "gatsby develop"
+    "build": "next build",
+    "dev": "next",
+    "start": "next start"
   },
   "dependencies": {
-    "@mdx-js/mdx": "^1.0.0",
-    "@mdx-js/react": "^1.0.0",
-    "gatsby": "^2.0.0",
-    "gatsby-plugin-mdx": "^1.0.0",
-    "gatsby-source-filesystem": "^2.0.0",
+    "@mdx-js/loader": "2.0.0-next.9",
+    "@mdx-js/mdx": "2.0.0-next.9",
+    "@next/mdx": "^10.0.0",
+    "next": "^11.0.0",
     "react": "^17.0.0",
     "react-dom": "^17.0.0"
   }
