+++ examples/react-web-components/pages/_app.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+const Page = ({Component}) => <Component />
+
+export default Page
