+++ examples/react-web-components/pages/index.mdx
@@ -1,0 +1,5 @@
+import {Counter} from './components/Counter.js'
+
+# Hello, world!
+
+<Counter />
