--- examples/react-web-components/src/components/Layout.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default ({children}) => (
-  <>
-    <main>{children}</main>
-  </>
-)
