--- examples/react-web-components/src/pages/index.mdx
@@ -1,5 +1,0 @@
-import WebComponent from '../components/WebComponent'
-
-# Hello, world!
-
-<WebComponent />
